Add Makefile and split up main
