+++ src/algebra.cpp
@@ -1,0 +1,32 @@
+#include <iostream>
+#include <string>
+using namespace std;
+
+#include "functions.h"
+
+int multiply(int a, int b)
+{
+	return a*b;
+}
+
+
+char mychar(char a, char b)
+{
+	cout << "first char is:";
+	cout << a << endl;
+	cout << "second = " ;
+	cout << b << endl;
+	return a+b;
+}
+
+string mystring(string a, string b)
+{
+	cout << a+b;
+	return b;
+}
+
+float myfrac(float a, float b)
+{
+	return a/b;
+}
+
