+++ src/functions.h
@@ -1,0 +1,11 @@
+// function defs
+//
+// #################################
+void helloworld();
+float myfrac(float a, float b);
+int multiply(int a, int b);
+char mychar(char a, char b);
+string mystring(string a, string b);
+//###################################
+
+
