+++ src/helloworld.cpp
@@ -1,0 +1,7 @@
+#include <iostream>
+using namespace std;
+
+void helloworld()
+{
+	cout << "Hello World!" <<endl;
+}
